Make createTestObjectProvider injectable
