+++ packages/test/end-to-end-tests/src/test/oldVersion.ts
@@ -34,13 +34,9 @@ import {
     TestContainerRuntimeFactory,
     TestFluidObjectFactory,
 } from "old-test-utils";
-import { Loader } from "old-container-loader";
 
 import {
-    createRuntimeFactory,
     DataObjectFactoryType,
-    getDataStoreFactory,
-    ITestObjectProvider,
     ITestContainerConfig,
     V1,
     V2,
@@ -114,7 +110,7 @@ const createOldTestFluidDataStoreFactory = (
     return new TestFluidObjectFactory(convertRegistry(registry));
 };
 
-function getOldDataStoreFactory(containerOptions?: ITestContainerConfig) {
+export function getDataStoreFactory(containerOptions?: ITestContainerConfig) {
     switch (containerOptions?.fluidDataObjectType) {
         case undefined:
         case DataObjectFactoryType.Primed:
@@ -126,7 +122,7 @@ function getOldDataStoreFactory(containerOptions?: ITestContainerConfig) {
     }
 }
 
-const createOldTestRuntimeFactory = (
+export const createRuntimeFactory = (
     type: string,
     dataStoreFactory: newVer.IFluidDataStoreFactory | IFluidDataStoreFactory,
     runtimeOptions: IContainerRuntimeOptions = { initialSummarizerDelayMs: 0 },
@@ -141,32 +137,4 @@ export function createOldRuntimeFactory(dataStore): IRuntimeFactory {
         factory,
         [[type, Promise.resolve(new DataObjectFactory(type, dataStore, [], {}))]],
     );
-}
-
-export function createTestObjectProvider(
-    oldLoader: boolean,
-    oldContainerRuntime: boolean,
-    oldDataStoreRuntime: boolean,
-    type: string,
-    serviceConfiguration?: Partial<newVer.IClientConfiguration>,
-    driver?: newVer.ITestDriver,
-): ITestObjectProvider {
-    const containerFactoryFn = (containerOptions?: ITestContainerConfig) => {
-        const dataStoreFactory = oldDataStoreRuntime
-            ? getOldDataStoreFactory(containerOptions)
-            : getDataStoreFactory(containerOptions);
-
-        return oldContainerRuntime
-            ? createOldTestRuntimeFactory(type, dataStoreFactory, containerOptions?.runtimeOptions)
-            : createRuntimeFactory(type, dataStoreFactory, containerOptions?.runtimeOptions);
-    };
-
-    if (driver === undefined) {
-        throw new Error("Must provide a driver when using the current loader");
-    }
-
-    return new newVer.TestObjectProvider(
-        oldLoader ? Loader as unknown as typeof newVer.Loader : newVer.Loader,
-        driver,
-        containerFactoryFn as () => newVer.IRuntimeFactory);
 }
